requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autogluon.common==1.1.1b20240426
autogluon.core==1.1.1b20240426
autogluon.features==1.1.1b20240426
-e git+https://github.com/autogluon/autogluon@87762b7bd2866abcce39f2793df7626c9787461e#egg=autogluon.multimodal&subdirectory=multimodal
babel==2.16.0
beautifulsoup4==4.12.3
black==23.12.1
bleach==6.1.0
boto3==1.34.92
botocore==1.34.92
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
comm==0.2.2
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
datasets==2.14.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.8
evaluate==0.4.1
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
future==1.0.0
gdown==5.1.0
google-api-core==2.18.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
hyperopt==0.2.7
idna==3.7
imageio==2.34.1
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.0
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
lazy_loader==0.4
lightning==2.1.4
lightning-utilities==0.11.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
model-index==0.1.11
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nlpaug==1.1.11
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
nptyping==2.4.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.2.3
onnx==1.13.1
onnxruntime-gpu==1.15.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.38
ordered-set==4.1.0
oss2==2.17.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==2.2.3
pytorch-metric-learning==2.3.0
pytz==2023.4
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.1.1
ray==2.10.0
referencing==0.35.0
regex==2024.4.16
requests==2.32.3
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-image==0.20.0
scikit-learn==1.4.0
scipy==1.12.0
Send2Trash==1.8.3
sentencepiece==0.2.0
seqeval==1.2.2
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.4.0
tifffile==2024.4.24
timm==0.9.16
tinycss2==1.3.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.1.2
torchmetrics==1.2.1
torchvision==0.16.2
tornado==6.4.1
tqdm==4.65.2
traitlets==5.14.3
transformers==4.38.2
triton==2.1.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.18
virtualenv==20.26.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
